engineering-thesis-django/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cffi==1.16.0
cryptography==41.0.4
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
packaging==23.2
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
